+++ src/main.cpp
@@ -335,7 +335,7 @@ int noclip(CommandLine& cli) {
 	if (cli.hasOption("-model")) {
 		model = cli.getOptionInt("-model");
 
-		if (model < 0 || model >= map->modelCount) {
+		if (model < 0 || (unsigned int)model >= map->modelCount) {
 			logf("ERROR: model number must be 0 - %d\n", map->modelCount);
 			return 1;
 		}
@@ -424,7 +424,7 @@ int simplify(CommandLine& cli) {
 
 	STRUCTCOUNT oldCounts(map);
 
-	if (modelIdx < 0 || modelIdx >= map->modelCount) {
+	if (modelIdx < 0 || (unsigned int)modelIdx >= map->modelCount) {
 		logf("ERROR: model number must be 0 - %d\n", map->modelCount);
 		return 1;
 	}
